Spread some ruby love
This commit is highly opiniated. There is no behavior change, but
a lot of changes to adjust the code more towards the usual ruby style
(or at least how I like to see ruby code). Nevertheless it also
addresses some pitfalls that are common when starting with ruby. The
points addressed in this commit are:

* single if not -> use unless instead
* no need for `then` after an if statement
* use the full power of enumerable: if arr.find_all{...}.empty? ->
  arr.none?{ ... }
* Array(foo).flatten -> [*foo]
* Pay attention to `and` & `or`, these are not the same as `&&` or `||`
  as most people think.
* [a].flatten! does not flatten `a`, it flattens [a] which is not
  assigned to anything.
* No need for a return statement at the end of a method
* + Some other details.

This is work is not a completely finished and as I stated some changes
are highly opiniated. But for me these are some necessary steps to work
further on that module.